requirements.txt
web: gunicorn config.wsgi --log-file -asgiref==3.3.1
beautifulsoup4==4.9.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.1.7
django-bootstrap3==14.2.0
django-bootstrap5==0.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.13.1
et-xmlfile==1.1.0
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
gspread==5.4.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
iso3166==2.0.2
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.5.2.54
openpyxl==3.0.10
pandas==1.4.2
pi==0.1.2
pi3d==2.45
Pillow==8.2.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.9
wget==3.2
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
psycopg2==2.8.6